Refactor Zoom Room management out of AdminController
The AdminController was getting entirely unwieldy. So let's start reworking things to keep logical concepts together.
