+++ arisia-remote/app/arisia/controllers/ZoomController.scala
@@ -1,11 +1,16 @@
 package arisia.controllers
 
-import arisia.models.{LoginUser, ProgramItemId}
+import arisia.admin.RoomService
+import arisia.auth.LoginService
+import arisia.models.{ProgramItemId, LoginUser, ZoomRoom}
 import arisia.schedule.ScheduleService
 import arisia.zoom.ZoomService
+import play.api.data._
+import play.api.data.Forms._
 import play.api.mvc.{BaseController, ControllerComponents, EssentialAction}
+import play.api.i18n.I18nSupport
 
-import scala.concurrent.ExecutionContext
+import scala.concurrent.{Future, ExecutionContext}
 
 /**
  * This controller manages all external entry points that are Zoom-specific.
@@ -13,10 +18,15 @@ import scala.concurrent.ExecutionContext
 class ZoomController(
   val controllerComponents: ControllerComponents,
   zoomService: ZoomService,
-  scheduleService: ScheduleService
+  roomService: RoomService,
+  scheduleService: ScheduleService,
+  val loginService: LoginService
 )(
-  implicit ec: ExecutionContext
-) extends BaseController {
+  implicit val ec: ExecutionContext
+) extends BaseController
+  with AdminControllerFuncs
+  with I18nSupport
+{
 
   // TODO: get rid of this. There isn't much harm to it, but we won't need it once integration is really working,
   // and there is no point leaving extra entry points around for potential mischief.
@@ -53,4 +63,66 @@ class ZoomController(
     // to get into this state, but we should probably provide a better error.
     redirectOpt.getOrElse(NotFound("""{"success":false, "message":"That isn't a currently-running panel"}"""))
   }
+
+  /* ********************
+   *
+   * Zoom Room CRUD
+   */
+
+  val roomForm = Form(
+    mapping(
+      "id" -> number,
+      "displayName" -> nonEmptyText,
+      "zoomId" -> nonEmptyText,
+      "zambiaName" -> nonEmptyText,
+      "isManual" -> boolean,
+      "isWebinar" -> boolean
+    )(ZoomRoom.apply)(ZoomRoom.unapply)
+  )
+
+  def manageZoomRooms(): EssentialAction = superAdminsOnly { info =>
+    implicit val request = info.request
+
+    val rooms = roomService.getRooms()
+    Ok(arisia.views.html.manageZoomRooms(rooms))
+  }
+
+  def createRoom(): EssentialAction = superAdminsOnly { info =>
+    implicit val request = info.request
+
+    Ok(arisia.views.html.editRoom(roomForm.fill(ZoomRoom.empty)))
+  }
+  def showEditRoom(id: Int): EssentialAction = superAdminsOnly { info =>
+    implicit val request = info.request
+
+    val rooms = roomService.getRooms()
+    rooms.find(_.id == id) match {
+      case Some(room) => Ok(arisia.views.html.editRoom(roomForm.fill(room)))
+      case _ => BadRequest(s"$id isn't a known Room!")
+    }
+  }
+
+  def roomModified(): EssentialAction = superAdminsOnlyAsync { info =>
+    implicit val request = info.request
+
+    roomForm.bindFromRequest().fold(
+      formWithErrors => {
+        // TODO: actually display the error!
+        Future.successful(BadRequest(arisia.views.html.editRoom(formWithErrors)))
+      },
+      room => {
+        val fut =
+          if (room.id == 0) {
+            roomService.addRoom(room)
+          } else {
+            roomService.editRoom(room)
+          }
+
+        fut.map { _ =>
+          Redirect(arisia.controllers.routes.ZoomController.manageZoomRooms())
+        }
+      }
+    )
+  }
+
 }
